+++ src/pyrosimple/scripts/rtcontrol.py
@@ -39,6 +39,18 @@ def print_help_fields():
         "call an arbiratry item method, e.g. 'd_session_file'"
         return ("d_METHOD", d_manifold)
 
+    def p_manifold():
+        "call an arbiratry p.multicall method, e.g. 'p_is_incoming'"
+        return ("p_METHOD", p_manifold)
+
+    def f_manifold():
+        "call an arbiratry f.multicall method, e.g. 'f_path'"
+        return ("f_METHOD", f_manifold)
+
+    def t_manifold():
+        "call an arbiratry t.multicall method, e.g. 't_url'"
+        return ("t_METHOD", t_manifold)
+
     print("")
     print("Fields are:")
     print(
@@ -51,6 +63,9 @@ def print_help_fields():
                         custom_manifold(),
                         kind_manifold(),
                         d_manifold(),
+                        t_manifold(),
+                        p_manifold(),
+                        f_manifold(),
                     ]
                 )
             ]